+++ crypt/test/odes/TestMirams2010WntOdeSystem.hpp
@@ -77,9 +77,9 @@ public:
         std::vector<double> initial_conditions = wnt_system.GetInitialConditions();
 
         double start_time, end_time, elapsed_time = 0.0;
-        start_time = std::clock();
+        start_time = (double) std::clock();
         solutions = cvode_solver.Solve(&wnt_system, initial_conditions, 0.0, 100.0, h_value, h_value);
-        end_time = std::clock();
+        end_time = (double) std::clock();
         elapsed_time = (end_time - start_time)/(CLOCKS_PER_SEC);
         std::cout << "1. Cvode Elapsed time = " << elapsed_time << " secs for 100 hours\n";
 
@@ -114,9 +114,9 @@ public:
         std::vector<double> initial_conditions = wnt_system.GetInitialConditions();
 
         double start_time, end_time, elapsed_time = 0.0;
-        start_time = std::clock();
+        start_time = (double) std::clock();
         solutions = cvode_solver.Solve(&wnt_system, initial_conditions, 0.0, 100.0, h_value, h_value);
-        end_time = std::clock();
+        end_time = (double) std::clock();
         elapsed_time = (end_time - start_time)/(CLOCKS_PER_SEC);
         std::cout << "1. Cvode Elapsed time = " << elapsed_time << " secs for 100 hours\n";
 
@@ -152,9 +152,9 @@ public:
         std::vector<double> initial_conditions = wnt_system.GetInitialConditions();
 
         double start_time, end_time, elapsed_time = 0.0;
-        start_time = std::clock();
+        start_time = (double) std::clock();
         solutions = cvode_solver.Solve(&wnt_system, initial_conditions, 0.0, 100.0, h_value, h_value);
-        end_time = std::clock();
+        end_time = (double) std::clock();
         elapsed_time = (end_time - start_time)/(CLOCKS_PER_SEC);
         std::cout << "1. Cvode Elapsed time = " << elapsed_time << " secs for 100 hours\n";
