+++ crypt/test/simulation/TestCryptSimulation2dWithMeshBasedCellPopulation.hpp
@@ -116,12 +116,12 @@ private:
     double mLastStartTime;
     void setUp()
     {
-        mLastStartTime = std::clock();
+        mLastStartTime = (double) std::clock();
         AbstractCellBasedTestSuite::setUp();
     }
     void tearDown()
     {
-        double time = std::clock();
+        double time = (double) std::clock();
         double elapsed_time = (time - mLastStartTime)/(CLOCKS_PER_SEC);
         std::cout << "Elapsed time: " << elapsed_time << std::endl;
         AbstractCellBasedTestSuite::tearDown();
